// types/global.d.ts
export type BotStats = {
  commands:  Record<string, number>
  users:     Record<string, number>
  hours:     Record<string, number>
  days:      Record<string, number>
  total:     number
  startedAt: number
}